Make financial inputs accept decimal values for accuracy
Update PlanForm.tsx to use parseFloat for monthly income, expenses, and savings inputs to correctly handle decimal values, and add console logs to server/routes.ts for better debugging of scenario and income item creation.

+++ client/src/pages/plan-form.tsx
@@ -315,7 +315,7 @@ export default function PlanForm() {
                               className="pl-8"
                               {...field}
                               value={field.value || ''}
-                              onChange={(e) => field.onChange(e.target.value ? parseInt(e.target.value) : undefined)}
+                              onChange={(e) => field.onChange(e.target.value ? parseFloat(e.target.value) : undefined)}
                               data-testid="input-monthly-income"
                             />
                           </div>
@@ -342,7 +342,7 @@ export default function PlanForm() {
                               className="pl-8"
                               {...field}
                               value={field.value || ''}
-                              onChange={(e) => field.onChange(e.target.value ? parseInt(e.target.value) : undefined)}
+                              onChange={(e) => field.onChange(e.target.value ? parseFloat(e.target.value) : undefined)}
                               data-testid="input-monthly-expenses"
                             />
                           </div>
@@ -369,7 +369,7 @@ export default function PlanForm() {
                               className="pl-8"
                               {...field}
                               value={field.value || ''}
-                              onChange={(e) => field.onChange(e.target.value ? parseInt(e.target.value) : undefined)}
+                              onChange={(e) => field.onChange(e.target.value ? parseFloat(e.target.value) : undefined)}
                               data-testid="input-monthly-savings"
                             />
                           </div>
